requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
pillow==9.4.0
torch==2.0.1
torchvision==0.15.2
